+++ lib/components/json_store/local_app_config.dart
@@ -2,7 +2,7 @@ import 'json_store.dart';
 
 class LocalAppConfig extends JsonStore {
   LocalAppConfig(String fp, {Duration? lapse})
-      : super(fp, lapse: lapse, indent: '  ');
+      : super.raw(fp, lapse: lapse, indent: '  ');
 
   JsonStoreItem<bool> get alwaysOnTop =>
       JsonStoreItem<bool>(this, 'alwaysOnTop');
@@ -12,4 +12,12 @@ class LocalAppConfig extends JsonStore {
   JsonStoreItem<int> get ffoSort => JsonStoreItem<int>(this, 'ffoSort');
 
   JsonStoreItem<int> get launchTimes => JsonStoreItem<int>(this, 'launchTimes');
+}
+
+class PersistentAppConfig extends JsonStore {
+  PersistentAppConfig(String fp)
+      : super.raw(fp, indent: '  ', lapse: Duration.zero);
+
+  JsonStoreItem get androidUseExternal =>
+      JsonStoreItem(this, 'android_use_external');
 }