+++ lib/components/datatypes/user/userdata.dart
@@ -86,6 +86,7 @@ class UserData {
 
   void validate() {
     if (db.gameData.servants.isNotEmpty) {
+      // TODO: if some servants missed in dataset, these data will be removed
       curUser.servants.removeWhere((key, value) =>
       db.gameData.unavailableSvts.contains(key) ||
           db.gameData.servantsWithUser[key] == null);